src/requirements.txt
# Core AI orchestration
crewai>=0.1.0
crewai[tools]

# LLM providers
openai>=0.27.0
anthropic>=0.3.0

# Configuration & YAML parsing
python-dotenv>=0.21.0
PyYAML>=6.0

# Optional code parsing/helper libraries
javalang>=0.13.0

# eventually to add testing
pytest>=7.2.0

mdutils

torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
transformers==4.51.3
langtrace-python-sdk
langchain-community